Fix access modifiers and move methods to improve readability
